requirements.txt
pyjwt
flask
flask-sqlalchemy
sqlalchemy
flask-restful
passlib
uuid
pyodbc
flasgger
gevent
